Fix a lot of bugs (#10740)
Fix #10736, #10688, #10724, #10696, #10687, #10651, #10586, #10498, #10466, #10433, #10454, #10319, #10366

* Fix a bug in influence cost calculation that could cause distant plots to be cheaper than near ones

* All quests end on, not after their end turn

* Fix not being able to buy units in automated cities

* Increase number of modifiers shown in enemy unit panel

* Fix units not being shown in city view when unit number limit reached

* Fix CTD caused by modders using negative values for GreatWorkCount

* Added check for IsDummy in CityView.lua

* Fix incorrect deal history when a player is killed

* Fix CTD when a city with an active event choice is no longer owned

* Fix corporations not being destroyed correctly

* Fix spaceship parts being buildable when SV is turned off

* Fix unit teleporting

* Revert military might calculation to vanilla

* Fix atomic bombs being counted as civilian units

* Fix grand strategy priorities being calculated for minor civs

+++ CvGameCoreDLL_Expansion2/CvPlot.cpp
@@ -649,7 +649,7 @@ void CvPlot::updateCenterUnit()
 
 
 //	--------------------------------------------------------------------------------
-void CvPlot::verifyUnitValidPlot(bool bWakeUp)
+void CvPlot::verifyUnitValidPlot(PlayerTypes eForSpecificPlayer, bool bWakeUp)
 {
 	vector<IDInfo> oldUnitList;
 
@@ -665,23 +665,26 @@ void CvPlot::verifyUnitValidPlot(bool bWakeUp)
 		CvUnit* pLoopUnit = GetPlayerUnit(oldUnitList[iVectorLoop]);
 		if(pLoopUnit != NULL)
 		{
-			if(!pLoopUnit->isDelayedDeath())
+			if (eForSpecificPlayer == NO_PLAYER || (eForSpecificPlayer != NO_PLAYER && pLoopUnit->getOwner() == eForSpecificPlayer))
 			{
-				if(pLoopUnit->atPlot(*this))
+				if (!pLoopUnit->isDelayedDeath())
 				{
-					//if plot ownership changes we want to make sure that human units
-					//which would suffer attrition or cause annoyance are woken up
-					if (bWakeUp)
-						pLoopUnit->SetActivityType(ACTIVITY_AWAKE);
-
-					if(!(pLoopUnit->isCargo()))
+					if (pLoopUnit->atPlot(*this))
 					{
-						if(!(pLoopUnit->isInCombat()))
+						//if plot ownership changes we want to make sure that human units
+						//which would suffer attrition or cause annoyance are woken up
+						if (bWakeUp)
+							pLoopUnit->SetActivityType(ACTIVITY_AWAKE);
+
+						if (!(pLoopUnit->isCargo()))
 						{
-							if (!pLoopUnit->canEndTurnAtPlot(this))
+							if (!(pLoopUnit->isInCombat()))
 							{
-								if (!pLoopUnit->jumpToNearestValidPlot())
-									pLoopUnit->kill(true);
+								if (!pLoopUnit->canEndTurnAtPlot(this))
+								{
+									if (!pLoopUnit->jumpToNearestValidPlot())
+										pLoopUnit->kill(true);
+								}
 							}
 						}
 					}
@@ -6764,7 +6767,7 @@ void CvPlot::setOwner(PlayerTypes eNewValue, int iAcquiringCityID, bool bCheckUn
 
 			if(bCheckUnits)
 			{
-				verifyUnitValidPlot(true);
+				verifyUnitValidPlot(NO_PLAYER, true);
 			}
 
 			if(GC.getGame().isDebugMode())
